feat: clean up downloadByRange
